requirements.txt
argparse==1.2.1
Babel==1.3
bcrypt==1.0.2
blinker==1.3
cffi==0.8.2
dnspython==1.11.1
easyzone==1.2.2
fess-up==1.1.1
Flask==0.10.1
Flask-Admin==1.0.8
Flask-Babel==0.9
Flask-Cache==0.13.1
Flask-Login==0.2.11
Flask-Mail==0.9.0
Flask-WTF==0.9.5
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
MySQL-python==1.2.5
psutil==2.1.3
py-bcrypt==0.4
pycparser==2.10
python-gnupg==0.3.6
python-redmine==0.8.2
python-whois==0.3.1
pytz==2014.9
PyYAML==3.11
requests==2.3.0
speaklater==1.3
Werkzeug==0.9.6
wheel==0.24.0
WTForms==1.0.5
